pyproject.toml
[project]
name = "linkedin-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "haystack-ai>=2.9.0",
    "linkedin-api>=2.3.1",
    "md2pdf>=1.0.1",
    "mistral-haystack>=0.0.2",
    "mistralai>=1.3.1",
    "pydantic>=2.10.5",
    "pydantic-settings>=2.7.1",
    "trafilatura[all]>=2.0.0",
    "unstructured-client>=0.29.0",
    "weaviate-client>=4.10.4",
    "google-api-python-client>=2.159.0",
    "google-auth>=2.37.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
]
classifiers = ["Private :: Do Not Upload"]

[dependency-groups]
dev = [
    "jupyterlab>=4.3.4",
    "pyright>=1.1.392.post0",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.2",
    "tqdm>=4.67.1",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "gql",
]

# Assume Python 3.12
target-version = "py312"

lint.select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

lint.ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]

[tool.ruff.lint.isort]
known-third-party = ["pydantic"]

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportMissingImports = "error"
reportMissingTypeStubs = false
exclude = ["**/__pycache__"]
include = ["./src"]
